requirements.txt
python-dotenv>=1.0.0
langchain-core>=0.2.0
langchain-google-community>=0.1.0
langchain-openai>=0.1.0
fastmcp==2.11.0
